log10 + double the avgdl

+++ postingList.h
@@ -22,14 +22,14 @@ class PostingList {
 	//Total times the word was found
 	int totalTimes;
 	double idf(const int& n, const int& nqi);
-	double fscore(const double& idf, const int& tf, const int& avgdl, const int& D, const double& k1 = 1.2, const double& b = 0.75);
+	double fscore(const double& idf, const int& tf, const double& avgdl, const int& D, const double& k1 = 1.2, const double& b = 0.75);
 public:
 	PostingList(int line);
 	void add(int line);
 	//Number of times a word appears in the file, if needed in future problem
 	int getTotalTimes();
 	int countNodes();
-	void score(double* bm25, bool* flags, int avgdl, int N, int* nwords);
+	void score(double* bm25, bool* flags, double avgdl, int N, int* nwords);
 	//term frequency function
 	int tf(int line);
 	~PostingList();
